Update Redis integration tests to use ConsoleMF
Also updated testing version to latest
