+++ tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/MultiFunctionApplicationHelpers.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFrameworks>netcoreapp3.1;net5.0;net6.0;net462;net471;net48</TargetFrameworks>
@@ -50,6 +50,16 @@
     <PackageReference Include ="MySqlConnector" Version="1.0.1" Condition="'$(TargetFramework)' == 'netcoreapp3.1'" />
     <PackageReference Include ="MySqlConnector" Version="1.3.13" Condition="'$(TargetFramework)' == 'net5.0'" />
     <PackageReference Include ="MySqlConnector" Version="2.1.2" Condition="'$(TargetFramework)' == 'net6.0'" />
+
+    <!-- StackExchange.Redis framework references -->
+    <PackageReference Include ="StackExchange.Redis.StrongName" Version="1.1.608" Condition="'$(TargetFramework)' == 'net462'" />
+    <PackageReference Include ="StackExchange.Redis" Version="2.0.601" Condition="'$(TargetFramework)' == 'net471'" />
+    <PackageReference Include ="StackExchange.Redis" Version="2.2.88" Condition="'$(TargetFramework)' == 'net48'" />
+    
+    <!-- StackExchange.Redis .NET/Core references -->
+    <PackageReference Include ="StackExchange.Redis.StrongName" Version="1.2.6" Condition="'$(TargetFramework)' == 'netcoreapp3.1'" />
+    <PackageReference Include ="StackExchange.Redis" Version="2.1.58" Condition="'$(TargetFramework)' == 'net5.0'" />
+    <PackageReference Include ="StackExchange.Redis" Version="2.6.66" Condition="'$(TargetFramework)' == 'net6.0'" />
   
     <!-- Serilog .NET framework references -->
     <PackageReference Include="Serilog" Version="2.0.0" Condition="'$(TargetFramework)' == 'net462'" />
