+++ tests/Agent/IntegrationTests/UnboundedApplications/BasicMvcApplication/BasicMvcApplication.csproj
@@ -156,7 +156,6 @@
     <Compile Include="Controllers\PostgresController.cs" />
     <Compile Include="Controllers\OracleController.cs" />
     <Compile Include="Controllers\MsSqlController.cs" />
-    <Compile Include="Controllers\RedisController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
