+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Redis/StackExchangeRedisTests.cs
@@ -1,7 +1,6 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-
 using NewRelic.Agent.IntegrationTestHelpers;
 using NewRelic.Agent.IntegrationTestHelpers.Models;
 using NewRelic.Agent.IntegrationTests.Shared;
@@ -11,18 +10,22 @@ using System.Collections.Generic;
 using System.Linq;
 using Xunit;
 using Xunit.Abstractions;
+using MultiFunctionApplicationHelpers;
 
 namespace NewRelic.Agent.UnboundedIntegrationTests.Redis
 {
-    [NetFrameworkTest]
-    public class StackExchangeRedisTests : NewRelicIntegrationTest<RemoteServiceFixtures.BasicMvcApplication>
+    public abstract class StackExchangeRedisTestsBase<TFixture> : NewRelicIntegrationTest<TFixture> where TFixture : ConsoleDynamicMethodFixture
     {
-        private readonly RemoteServiceFixtures.BasicMvcApplication _fixture;
+        private readonly ConsoleDynamicMethodFixture _fixture;
 
-        public StackExchangeRedisTests(RemoteServiceFixtures.BasicMvcApplication fixture, ITestOutputHelper output)  : base(fixture)
+        public StackExchangeRedisTestsBase(TFixture fixture, ITestOutputHelper output) : base(fixture)
         {
             _fixture = fixture;
             _fixture.TestLogger = output;
+
+            _fixture.AddCommand($"StackExchangeRedisExerciser DoSomeWork");
+            _fixture.AddCommand($"StackExchangeRedisExerciser DoSomeWorkAsync");
+
             _fixture.Actions
             (
                 setupConfiguration: () =>
@@ -30,18 +33,17 @@ namespace NewRelic.Agent.UnboundedIntegrationTests.Redis
                     var configPath = fixture.DestinationNewRelicConfigFilePath;
                     var configModifier = new NewRelicConfigModifier(configPath);
 
-                    configModifier.ForceTransactionTraces();
+                    configModifier.ForceTransactionTraces()
+                    .SetLogLevel("finest");
 
                     var instrumentationFilePath = $@"{fixture.DestinationNewRelicExtensionsDirectoryPath}\NewRelic.Providers.Wrapper.Sql.Instrumentation.xml";
                     CommonUtils.SetAttributeOnTracerFactoryInNewRelicInstrumentation(
                         instrumentationFilePath, "NewRelic.Agent.Core.Tracer.Factories.Sql.DataReaderTracerFactory", "enabled", "true");
-                },
-                exerciseApplication: () =>
-                {
-                    _fixture.GetStackExchangeRedis();
-                    _fixture.GetStackExchangeRedisAsync();
                 }
             );
+
+            // Confirm both transaction transforms have completed before moving on to host application shutdown, and final sendDataOnExit harvest
+            _fixture.AddActions(exerciseApplication: () => _fixture.AgentLog.WaitForLogLines(AgentLogBase.TransactionTransformCompletedLogLineRegex, TimeSpan.FromMinutes(2), 2));
 
             _fixture.Initialize();
         }
@@ -49,12 +51,15 @@ namespace NewRelic.Agent.UnboundedIntegrationTests.Redis
         [Fact]
         public void Test()
         {
+            var syncTransactionName = "OtherTransaction/Custom/MultiFunctionApplicationHelpers.NetStandardLibraries.StackExchangeRedis.StackExchangeRedisExerciser/DoSomeWork";
+            var asyncTransactionName = "OtherTransaction/Custom/MultiFunctionApplicationHelpers.NetStandardLibraries.StackExchangeRedis.StackExchangeRedisExerciser/DoSomeWorkAsync";
+
             var expectedMetrics = new List<Assertions.ExpectedMetric>
             {
                 new Assertions.ExpectedMetric { metricName = @"Datastore/all", callCount = 62 },
-                new Assertions.ExpectedMetric { metricName = @"Datastore/allWeb", callCount = 62 },
+                new Assertions.ExpectedMetric { metricName = @"Datastore/allOther", callCount = 62 },
                 new Assertions.ExpectedMetric { metricName = @"Datastore/Redis/all", callCount = 62 },
-                new Assertions.ExpectedMetric { metricName = @"Datastore/Redis/allWeb", callCount = 62 },
+                new Assertions.ExpectedMetric { metricName = @"Datastore/Redis/allOther", callCount = 62 },
                 new Assertions.ExpectedMetric { metricName = $@"Datastore/instance/Redis/{CommonUtils.NormalizeHostname(StackExchangeRedisConfiguration.StackExchangeRedisServer)}/{StackExchangeRedisConfiguration.StackExchangeRedisPort}", callCount = 62},
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/SET", callCount = 2 },
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/SET", callCount = 2 },
@@ -75,7 +80,8 @@ namespace NewRelic.Agent.UnboundedIntegrationTests.Redis
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/EXISTS", callCount = 2 },
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/RANDOMKEY", callCount = 2 },
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/RENAME", callCount = 2 },
-                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/DEL", callCount = 2 },
+                // Delete can resolve to DEL or UNLINK depending on Redis version
+                new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/(DEL|UNLINK)", IsRegexName = true, callCount = 2 },
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/PING", callCount = 4 }, //ping and identifyendpoint
 				new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/SADD", callCount = 4 },
                 new Assertions.ExpectedMetric { metricName = @"Datastore/operation/Redis/SUNION", callCount = 2 },
@@ -91,9 +97,9 @@ namespace NewRelic.Agent.UnboundedIntegrationTests.Redis
 
             var unexpectedMetrics = new List<Assertions.ExpectedMetric>
             {
-				// The datastore operation happened inside a web transaction so there should be no allOther metrics
-				new Assertions.ExpectedMetric {metricName = @"Datastore/allOther", callCount = 1},
-                new Assertions.ExpectedMetric {metricName = @"Datastore/Redis/allOther", callCount = 1}
+				// The datastore operation happened inside a console app so there should be no allWeb metrics
+				new Assertions.ExpectedMetric {metricName = @"Datastore/allWeb", callCount = 1},
+                new Assertions.ExpectedMetric {metricName = @"Datastore/Redis/allWeb", callCount = 1}
             };
 
             var expectedTransactionEventIntrinsicAttributes = new List<string>
@@ -102,21 +108,80 @@ namespace NewRelic.Agent.UnboundedIntegrationTests.Redis
             };
 
             var metrics = _fixture.AgentLog.GetMetrics().ToList();
-            var transactionSample = _fixture.AgentLog.TryGetTransactionSample("WebTransaction/MVC/RedisController/StackExchangeRedis");
-            var transactionEvent = _fixture.AgentLog.TryGetTransactionEvent("WebTransaction/MVC/RedisController/StackExchangeRedis");
+            // The previous version of this test cheated a bit by naming both transactions the same. Since either one might be the slowest,
+            // let's allow for either one
+            var transactionSample = _fixture.AgentLog.TryGetTransactionSample(syncTransactionName) ?? _fixture.AgentLog.TryGetTransactionSample(asyncTransactionName);
+            var transactionEvent1 = _fixture.AgentLog.TryGetTransactionEvent(syncTransactionName);
+            var transactionEvent2 = _fixture.AgentLog.TryGetTransactionEvent(asyncTransactionName);
 
             NrAssert.Multiple
             (
                 () => Assert.NotNull(transactionSample),
-                () => Assert.NotNull(transactionEvent)
+                () => Assert.NotNull(transactionEvent1),
+                () => Assert.NotNull(transactionEvent2)
             );
 
             NrAssert.Multiple
             (
                 () => Assertions.MetricsExist(expectedMetrics, metrics),
                 () => Assertions.MetricsDoNotExist(unexpectedMetrics, metrics),
-                () => Assertions.TransactionEventHasAttributes(expectedTransactionEventIntrinsicAttributes, TransactionEventAttributeType.Intrinsic, transactionEvent)
+                () => Assertions.TransactionEventHasAttributes(expectedTransactionEventIntrinsicAttributes, TransactionEventAttributeType.Intrinsic, transactionEvent1),
+                () => Assertions.TransactionEventHasAttributes(expectedTransactionEventIntrinsicAttributes, TransactionEventAttributeType.Intrinsic, transactionEvent2)
             );
+        }
+    }
+
+    [NetFrameworkTest]
+    public class StackExchangeRedisTestsFW462 : StackExchangeRedisTestsBase<ConsoleDynamicMethodFixtureFW462>
+    {
+        public StackExchangeRedisTestsFW462(ConsoleDynamicMethodFixtureFW462 fixture, ITestOutputHelper output) : base(fixture, output)
+        {
+
+        }
+    }
+
+    [NetFrameworkTest]
+    public class StackExchangeRedisTestsFW471 : StackExchangeRedisTestsBase<ConsoleDynamicMethodFixtureFW471>
+    {
+        public StackExchangeRedisTestsFW471(ConsoleDynamicMethodFixtureFW471 fixture, ITestOutputHelper output) : base(fixture, output)
+        {
+
+        }
+    }
+
+    [NetFrameworkTest]
+    public class StackExchangeRedisTestsFWLatest : StackExchangeRedisTestsBase<ConsoleDynamicMethodFixtureFWLatest>
+    {
+        public StackExchangeRedisTestsFWLatest(ConsoleDynamicMethodFixtureFWLatest fixture, ITestOutputHelper output) : base(fixture, output)
+        {
+
+        }
+    }
+
+    [NetCoreTest]
+    public class StackExchangeRedisTestsCore31 : StackExchangeRedisTestsBase<ConsoleDynamicMethodFixtureCore31>
+    {
+        public StackExchangeRedisTestsCore31(ConsoleDynamicMethodFixtureCore31 fixture, ITestOutputHelper output) : base(fixture, output)
+        {
+
+        }
+    }
+
+    [NetCoreTest]
+    public class StackExchangeRedisTestsCore50 : StackExchangeRedisTestsBase<ConsoleDynamicMethodFixtureCore50>
+    {
+        public StackExchangeRedisTestsCore50(ConsoleDynamicMethodFixtureCore50 fixture, ITestOutputHelper output) : base(fixture, output)
+        {
+
+        }
+    }
+
+    [NetCoreTest]
+    public class StackExchangeRedisTestsCoreLatest : StackExchangeRedisTestsBase<ConsoleDynamicMethodFixtureCoreLatest>
+    {
+        public StackExchangeRedisTestsCoreLatest(ConsoleDynamicMethodFixtureCoreLatest fixture, ITestOutputHelper output) : base(fixture, output)
+        {
+
         }
     }
 }
